Migrate remaining server functionality
