+++ commands/serve.js
@@ -1,31 +1,37 @@
 'use strict';
 
-// TODO: Remove gulp dependencies (ex. include and use `less` instead of `gulp-less`, remove `gulp`, etc)
 var
   fs = require('fs'),
 
   chalk    = require('chalk'),
   chokidar = require('chokidar'),
-  extend   = require('extend'),
+  open     = require('open'),
 
+  // TODO: Remove gulp dependencies (ex. include and use `less` instead of `gulp-less`, remove `gulp`, etc)
   gulp       = require('gulp'),
-  gutil      = require('gulp-util'),
   rename     = require('gulp-rename'),
-  replace    = require('gulp-replace'),
   less       = require('gulp-less'),
   handlebars = require('gulp-compile-handlebars'),
   connect    = require('gulp-connect'),
-  open       = require('gulp-open');
+
+  helpers = require('../lib/handlebars-helpers.js');
 
 var
-  css,
-  command,
   cwd,
+
   html,
+  css,
+  js,
+  hbs,
   server,
-  watch;
+  watch,
+  openUrl,
+
+  command;
+
 
 cwd = process.cwd();
+
 
 html = function() {
   gulp.src('*.html')
@@ -44,9 +50,34 @@ css = function() {
     .pipe(connect.reload());
 };
 
+js = function() {
+  gulp.src('*.js')
+    .pipe(connect.reload());
+};
+
+hbs = function() {
+  var
+    templateData,
+    options;
+
+  templateData = JSON.parse(fs.readFileSync('./customer.json', 'utf-8'));
+
+  options = {
+    helpers: helpers,
+    batch: ['templates']
+  };
+
+  gulp.src('templates/widget.hbs')
+    .pipe(handlebars(templateData, options))
+    .pipe(rename('index.html'))
+    .pipe(gulp.dest(''));
+};
+
 watch = function() {
-  chokidar.watch('*.html').on('all', html);
-  chokidar.watch('assets/css/*.less').on('all', css);
+  chokidar.watch('*.html').on('change', html);
+  chokidar.watch('assets/css/*.less').on('change', css);
+  chokidar.watch('assets/javascript/*.js').on('change', js);
+  chokidar.watch(['*.json', 'templates/*.hbs']).on('change', hbs);
 };
 
 server = function() {
@@ -57,6 +88,11 @@ server = function() {
   });
 };
 
+openUrl = function() {
+  open('http://0.0.0.0:8080');
+};
+
+
 command = function(program) {
   var serve;
 
@@ -65,9 +101,11 @@ command = function(program) {
   serve
     .description('Start a server for a theme')
     .action(function() {
-      css();
       server();
+      hbs();
+      css();
       watch();
+      openUrl();
     });
 
   return serve;
